chore: move /tests/unit/core to /tests/unit/service_layer
Test tree should follow the src tree
In fact we remove the /tests/unit/core folder
